miinaharava/pyproject.toml
[project]
name = "miinaharava"
version = "0.1.0"
description = ""
authors = [
    {name = "acpeltol",email = "apeltola717@gmial.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "invoke (>=2.2.0,<3.0.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pylint = "^3.3.6"
autopep8 = "^2.3.2"

